Refactoring scripts.
