+++ Assets/Scripts/Movement.cs
@@ -6,6 +6,8 @@ using UnityEngine;
 public class Movement : MonoBehaviour
 {
     [SerializeField] private float _speed = 5.0f;
+
+    private const string _horizontal = "Horizontal";
     
     private Rigidbody2D _rigidbody2D;
     private float _horizontalMove;
@@ -15,9 +17,7 @@ public class Movement : MonoBehaviour
 
     private void Update()
     {
-        _horizontalMove = Input.GetAxis("Horizontal") * _speed;
+        _horizontalMove = Input.GetAxis(_horizontal) * _speed;
         _rigidbody2D.velocity = new Vector2(_horizontalMove, _rigidbody2D.velocity.y);
     }
-    
-    
 }
