+++ Assets/Scripts/ThiefDetector.cs.meta
